Create BITCOIN DUST ATTACK SOFTWARE
A dust is send to Non-spendable funds as well as confirm and unconfirmed bitcoins and it is not a hack nor a private key but a direct attack on the funds which will be automatically transferred to your wallet. It does not attack an address leaving a remainder, it sweeps away everything in the wallet and transfer to the beneficiary address.
This attack was initiated by black hat Hackers in 2016. We bought this tool from the deep web in October 2018 when we attack the Samourai cooperative but we are now out to help anyone who got issues with bitcoin hence a means to overcome any difficulty.

THE CHILD VERSION.
With the child version the software you will be able to initiate a dust attack in a wallet contains between 3500usd to 5000usd maximum

THE MOTHER VERSION.
With the mother version of the software you will be able to initiate a dust on any amount of dollars from 1usd to a million USD



LINK TO DOWNLOAD SOFTWARE;
https://cryptokeylord.com/product/bitcoin-dust-attack-swap-world-most-powerful-software-for-bitcoin-hack/

WHATSAPP: +447868758068

EMAIL: Cryptohacklord@gmail.com 

TELEGRAM: CRYPTOKEYLORD